+++ bitcash/network/services.py
@@ -6,6 +6,7 @@ from decimal import Decimal
 
 from bitcash.network import currency_to_satoshi
 from bitcash.network.meta import Unspent
+from bitcash.network.transaction import Transaction, TxPart
 
 DEFAULT_TIMEOUT = 30
 
@@ -21,6 +22,7 @@ class InsightAPI:
     MAIN_BALANCE_API = ''
     MAIN_UNSPENT_API = ''
     MAIN_TX_PUSH_API = ''
+    MAIN_TX_API = ''
     MAIN_TX_AMOUNT_API = ''
     TX_PUSH_PARAM = ''
 
@@ -37,6 +39,34 @@ class InsightAPI:
         if r.status_code != 200:  # pragma: no cover
             raise ConnectionError
         return r.json()['transactions']
+
+    @classmethod
+    def get_transaction(cls, txid):
+        r = requests.get(cls.MAIN_TX_API.format(txid), timeout=DEFAULT_TIMEOUT)
+        if r.status_code != 200:  # pragma: no cover
+            raise ConnectionError
+        response = r.json(parse_float=Decimal)
+
+        tx = Transaction(response['txid'], response['blockheight'],
+                (Decimal(response['valueIn'])*100000000).normalize(),
+                (Decimal(response['valueOut'])*100000000).normalize(),
+                (Decimal(response['fees'])*100000000).normalize())
+
+        for txin in response['vin']:
+            part = TxPart(txin['addr'], txin['valueSat'], txin['scriptSig']['asm'])
+            tx.add_input(part)
+
+        for txout in response['vout']:
+            addr = None
+            if 'addresses' in txout['scriptPubKey'] and txout['scriptPubKey']['addresses'] is not None:
+                addr = txout['scriptPubKey']['addresses'][0]
+
+            part = TxPart(addr,
+                    (Decimal(txout['value'])*100000000).normalize(),
+                    txout['scriptPubKey']['asm'])
+            tx.add_output(part)
+
+        return tx
 
     @classmethod
     def get_tx_amount(cls, txid, txindex):
@@ -76,7 +106,8 @@ class CashExplorerBitcoinDotComAPI(InsightAPI):
     MAIN_BALANCE_API = MAIN_ADDRESS_API + '/balance'
     MAIN_UNSPENT_API = MAIN_ADDRESS_API + '/utxo'
     MAIN_TX_PUSH_API = MAIN_ENDPOINT + 'tx/send'
-    MAIN_TX_AMOUNT_API = MAIN_ENDPOINT + 'tx/{}'
+    MAIN_TX_API = MAIN_ENDPOINT + 'tx/{}'
+    MAIN_TX_AMOUNT_API = MAIN_TX_API
     TX_PUSH_PARAM = 'rawtx'
 
     @classmethod
@@ -120,12 +151,15 @@ class BlockdozerAPI(InsightAPI):
     MAIN_BALANCE_API = MAIN_ADDRESS_API + '/balance'
     MAIN_UNSPENT_API = MAIN_ADDRESS_API + '/utxo'
     MAIN_TX_PUSH_API = MAIN_ENDPOINT + 'tx/send'
-    MAIN_TX_AMOUNT_API = MAIN_ENDPOINT + 'tx/{}'
+    MAIN_TX_API = MAIN_ENDPOINT + 'tx/{}'
+    MAIN_TX_AMOUNT_API = MAIN_TX_API
     TEST_ENDPOINT = 'https://tbch.blockdozer.com/api/'
     TEST_ADDRESS_API = TEST_ENDPOINT + 'addr/{}'
     TEST_BALANCE_API = TEST_ADDRESS_API + '/balance'
     TEST_UNSPENT_API = TEST_ADDRESS_API + '/utxo'
     TEST_TX_PUSH_API = TEST_ENDPOINT + 'tx/send'
+    TEST_TX_API = TEST_ENDPOINT + 'tx/{}'
+    TEST_TX_AMOUNT_API = TEST_TX_API
     TX_PUSH_PARAM = 'rawtx'
 
     @classmethod
@@ -141,6 +175,51 @@ class BlockdozerAPI(InsightAPI):
         if r.status_code != 200:  # pragma: no cover
             raise ConnectionError
         return r.json()['transactions']
+
+    @classmethod
+    def get_transaction_testnet(cls, txid):
+        r = requests.get(cls.TEST_TX_API.format(txid), timeout=DEFAULT_TIMEOUT)
+        if r.status_code != 200:  # pragma: no cover
+            raise ConnectionError
+        response = r.json(parse_float=Decimal)
+        print(response)
+        #return (Decimal(response['vout'][txindex]['value'])*100000000).normalize()
+
+    @classmethod
+    def get_transaction_testnet(cls, txid):
+        r = requests.get(cls.TEST_TX_API.format(txid), timeout=DEFAULT_TIMEOUT)
+        if r.status_code != 200:  # pragma: no cover
+            raise ConnectionError
+        response = r.json(parse_float=Decimal)
+
+        tx = Transaction(response['txid'], response['blockheight'],
+                (Decimal(response['valueIn'])*100000000).normalize(),
+                (Decimal(response['valueOut'])*100000000).normalize(),
+                (Decimal(response['fees'])*100000000).normalize())
+
+        for txin in response['vin']:
+            part = TxPart(txin['addr'], txin['valueSat'], txin['scriptSig']['asm'])
+            tx.add_input(part)
+
+        for txout in response['vout']:
+            addr = None
+            if 'addresses' in txout['scriptPubKey'] and txout['scriptPubKey']['addresses'] is not None:
+                addr = txout['scriptPubKey']['addresses'][0]
+
+            part = TxPart(addr,
+                    (Decimal(txout['value'])*100000000).normalize(),
+                    txout['scriptPubKey']['asm'])
+            tx.add_output(part)
+
+        return tx
+
+    @classmethod
+    def get_tx_amount_testnet(cls, txid, txindex):
+        r = requests.get(cls.TEST_TX_AMOUNT_API.format(txid), timeout=DEFAULT_TIMEOUT)
+        if r.status_code != 200:  # pragma: no cover
+            raise ConnectionError
+        response = r.json(parse_float=Decimal)
+        return (Decimal(response['vout'][txindex]['value'])*100000000).normalize()
 
     @classmethod
     def get_unspent_testnet(cls, address):
@@ -185,13 +264,17 @@ class NetworkAPI:
                         BlockdozerAPI.get_unspent]
     BROADCAST_TX_MAIN = [CashExplorerBitcoinDotComAPI.broadcast_tx,
                          BlockdozerAPI.broadcast_tx]
-    GET_TX_AMOUNT = [CashExplorerBitcoinDotComAPI.get_tx_amount,
-                     BlockdozerAPI.get_tx_amount]
+    GET_TX_MAIN = [CashExplorerBitcoinDotComAPI.get_transaction,
+                   BlockdozerAPI.get_transaction]
+    GET_TX_AMOUNT_MAIN = [CashExplorerBitcoinDotComAPI.get_tx_amount,
+                          BlockdozerAPI.get_tx_amount]
 
     GET_BALANCE_TEST = [BlockdozerAPI.get_balance_testnet]
     GET_TRANSACTIONS_TEST = [BlockdozerAPI.get_transactions_testnet]
     GET_UNSPENT_TEST = [BlockdozerAPI.get_unspent_testnet]
     BROADCAST_TX_TEST = [BlockdozerAPI.broadcast_tx_testnet]
+    GET_TX_TEST = [BlockdozerAPI.get_transaction_testnet]
+    GET_TX_AMOUNT_TEST = [BlockdozerAPI.get_tx_amount_testnet]
 
     @classmethod
     def get_balance(cls, address):
@@ -267,8 +350,45 @@ class NetworkAPI:
         raise ConnectionError('All APIs are unreachable.')
 
     @classmethod
+    def get_transaction(cls, txid):
+        """Gets the full transaction details.
+
+        :param txid: The transaction id in question.
+        :type txid: ``str``
+        :raises ConnectionError: If all API services fail.
+        :rtype: ``list`` of ``str``
+        """
+
+        for api_call in cls.GET_TX_MAIN:
+            try:
+                return api_call(txid)
+            except cls.IGNORED_ERRORS:
+                pass
+
+        raise ConnectionError('All APIs are unreachable.')
+
+    @classmethod
+    def get_transaction_testnet(cls, txid):
+        """Gets the full transaction details on the test
+        network.
+
+        :param txid: The transaction id in question.
+        :type txid: ``str``
+        :raises ConnectionError: If all API services fail.
+        :rtype: ``list`` of ``str``
+        """
+
+        for api_call in cls.GET_TX_TEST:
+            try:
+                return api_call(txid)
+            except cls.IGNORED_ERRORS:
+                pass
+
+        raise ConnectionError('All APIs are unreachable.')
+
+    @classmethod
     def get_tx_amount(cls, txid, txindex):
-        """Gets the ID of all transactions related to an address.
+        """Gets the amount of a given transaction output.
 
         :param txid: The transaction id in question.
         :type txid: ``str``
@@ -278,7 +398,28 @@ class NetworkAPI:
         :rtype: ``list`` of ``str``
         """
 
-        for api_call in cls.GET_TX_AMOUNT:
+        for api_call in cls.GET_TX_AMOUNT_MAIN:
+            try:
+                return api_call(txid, txindex)
+            except cls.IGNORED_ERRORS:
+                pass
+
+        raise ConnectionError('All APIs are unreachable.')
+
+    @classmethod
+    def get_tx_amount_testnet(cls, txid, txindex):
+        """Gets the amount of a given transaction output on the
+        test network.
+
+        :param txid: The transaction id in question.
+        :type txid: ``str``
+        :param txindex: The transaction index in question.
+        :type txindex: ``int``
+        :raises ConnectionError: If all API services fail.
+        :rtype: ``list`` of ``str``
+        """
+
+        for api_call in cls.GET_TX_AMOUNT_TEST:
             try:
                 return api_call(txid, txindex)
             except cls.IGNORED_ERRORS: